requirements.txt
ultralytics==8.3.130
torch==2.6.0
transformers==4.51.3
matplotlib==3.10.3
opencv-python==4.11.0.86
gradio==5.29.0
Pillow==11.2.1
accelerate==1.6.0
bitsandbytes==0.45.5
autogen-agentchat==0.5.7
autogen-core==0.5.7
autogen-ext[openai,azure]==0.5.7
wikipedia==1.4.0
arxiv==2.2.0
scipy==1.15.2
langchain-community==0.3.24